requirements.txt
fastapi>=0.119.0
pydantic>=2.10.0
python-dotenv>=1.1.1
pyjwt>=2.10.1
uvicorn>=0.27.1
mysql-connector-python>=9.5.0
python-multipart>=0.0.9